chore: add Python service debugging config
Add debugpy to the Docker dev container for easier debugging of the Python
Bottle service. Launches the Python service with `python -m debug.py` to allow
attaching to the debugger at runtime. Note this will require a `launch.json`
config file to configure the debugger for VSCode to connect to the default port
(5678).

For example
```
{
  "version": "0.2.0",
    "configurations": [
        {
            "name": "Python Debugger: Attach using Process Id",
            "type": "debugpy",
            "request": "attach",
            "connect": {
                "host": "localhost",
                "port": 5678
            }
        }
    ]
}
```

Also adds a devcontainer config sample for Python and splits up the Laravel and
Python service configs, so they can be managed independently.